refactor(reward): persisted types into state.rs
