+++ actors/reward/src/streams/tests/mod.rs
@@ -4,9 +4,11 @@ use fvm_shared::econ::TokenAmount;
 use num_traits::Zero;
 
 use super::*;
+use crate::state::*;
 use crate::streams::award::Allocation;
 use crate::streams::invariants::schedule_at;
 use crate::streams::queue::Stranded;
+use crate::types::*;
 
 mod award;
 mod distribution;
